refactor: split server/api.ts into domain modules
Break up the 1,749-line server/api.ts monolith into focused modules
grouped by concern. No behavior changes — logic is preserved verbatim.

New layout under apps/web/server/:
- shared-types.ts       — SourceId, SocialLink, PlatformResult, etc.
- shared-utils.ts       — fetchWithTimeout, normalizeForComparison, parseReleaseDate, etc.
- social-links.ts       — Discogs/official-site social link scraping
- resolve-streaming-url.ts — Spotify/Apple Music URL → artist name
- search/               — one file per platform
    bandcamp.ts, bandwagon.ts, beatport.ts, even.ts, faircamp.ts,
    mirlo.ts, musicbrainz.ts, patreon.ts, qobuz.ts
    index.ts — searchAllPlatforms orchestrator + aggregateResults
- api.ts                — HTTP routing only (128 lines, down from 1,749)

Only handleApiRequest is consumed externally (by vite.config.ts);
module boundary is preserved.
