+++ apps/web/server/search/bandwagon.ts
@@ -1,0 +1,53 @@
+import { parse } from 'node-html-parser';
+import { fetchWithTimeout, normalizeForComparison } from '../shared-utils';
+
+// Search Bandwagon for artists by scraping search results.
+// Returns a map of normalized artist name -> direct Bandwagon artist URL.
+export async function searchBandwagon(query: string): Promise<Map<string, string>> {
+  const results = new Map<string, string>();
+  const searchUrl = `https://bandwagon.fm/artists?q=${encodeURIComponent(query)}`;
+
+  try {
+    const response = await fetchWithTimeout(searchUrl, {
+      headers: {
+        'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36',
+      },
+    }, 3000);
+
+    if (!response.ok) return results;
+
+    const html = await response.text();
+    const root = parse(html);
+    const queryNormalized = normalizeForComparison(query);
+
+    // Bandwagon uses /@{id} format
+    const artistLinks = root.querySelectorAll('a[href*="bandwagon.fm/@"]');
+    const seen = new Set<string>();
+
+    for (const link of artistLinks) {
+      const href = link.getAttribute('href');
+      const nameEl = link.querySelector('.bold');
+      const name = nameEl?.textContent?.trim();
+
+      if (href && name && !seen.has(href) && name.length > 0 && name.length < 100) {
+        seen.add(href);
+        const normalizedName = normalizeForComparison(name);
+
+        if (normalizedName === queryNormalized ||
+            normalizedName.includes(queryNormalized) ||
+            queryNormalized.includes(normalizedName)) {
+          if (!results.has(normalizedName)) {
+            results.set(normalizedName, href);
+          }
+          if (results.size >= 10) break;
+        }
+      }
+    }
+  } catch (error: any) {
+    if (error.name !== 'AbortError') {
+      console.error('Bandwagon search error:', error.message);
+    }
+  }
+
+  return results;
+}
