+++ apps/web/server/shared-types.ts
@@ -1,0 +1,61 @@
+export type SourceId =
+  | 'bandcamp'
+  | 'mirlo'
+  | 'ampwall'
+  | 'bandwagon'
+  | 'faircamp'
+  | 'patreon'
+  | 'buymeacoffee'
+  | 'kofi'
+  | 'hoopla'
+  | 'freegal'
+  | 'qobuz'
+  | 'beatport'
+  | 'even';
+
+export type SocialPlatform = 'instagram' | 'facebook' | 'tiktok' | 'youtube' | 'threads' | 'bluesky' | 'twitter';
+
+export interface SocialLink {
+  platform: SocialPlatform;
+  url: string;
+}
+
+export interface MusicBrainzEnrichmentResponse {
+  query: string;
+  artistName: string | null;
+  officialUrl: string | null;
+  discogsUrl: string | null;
+  hasPre2005Release: boolean;
+  socialLinks: SocialLink[];
+}
+
+export interface LatestRelease {
+  title: string;
+  type: 'album' | 'track';
+  url: string;
+  imageUrl?: string;
+  releaseDate?: string;
+}
+
+export interface PlatformResult {
+  sourceId: SourceId;
+  name: string;
+  artist?: string;
+  type: 'artist' | 'album' | 'track';
+  url: string;
+  imageUrl?: string;
+  latestRelease?: LatestRelease;
+}
+
+export interface AggregatedResult {
+  id: string;
+  name: string;
+  artist?: string;
+  type: 'artist' | 'album' | 'track';
+  imageUrl?: string;
+  platforms: {
+    sourceId: SourceId;
+    url: string;
+    latestRelease?: LatestRelease;
+  }[];
+}
